backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
openai>=1.10.0
google-generativeai>=0.3.2
python-docx>=1.1.0
openpyxl>=3.1.2
PyPDF2>=3.0.1
aiofiles>=23.2.1
sqlalchemy>=2.0.25
httpx>=0.26.0
